requirements.txt
click==6.7
curtsies==0.3.0
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Script==2.0.5
Flask-HTTPAuth==3.2.3
html5lib==0.999999999
gunicorn==19.7.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
rq==0.9.2
six==1.11.0
selenium==3.141.0
urllib3==1.24.1
visitor==0.1.3
telebot==0.0.3
pytelegrambotapi==3.6.7